tvShowsApp-back/tv_shows/requirements.txt
asgiref==3.7.2
Django==4.2.2
django-cors-headers==4.0.0
psycopg2-binary==2.9.6
sqlparse==0.4.4
typing_extensions==4.6.3